💥Remove a dumb scam reason
I see this as a really dumb reason for a ban of a friend of mine. It's really unfair to be excluded for advertising something that clearly was a joke. Your staff should honestly check more often what reasons are provided.
If you can please accept this pull request or at least check him again because I really feel this is unfair.
Thank you for getting through this.
Much appreciation for such a great community but this feels awful.